ui: refactor 'detailedResults' to not be an iframe
