+++ ui/core/components/detailed_results/aura_metrics.ts
@@ -1,5 +1,5 @@
-import { ActionId } from '../core/proto_utils/action_id';
-import { AuraMetrics, SimResult, SimResultFilter } from '../core/proto_utils/sim_result';
+import { ActionId } from '../../proto_utils/action_id';
+import { AuraMetrics, SimResult, SimResultFilter } from '../../proto_utils/sim_result';
 
 import { ColumnSortType, MetricsTable } from './metrics_table';
 import { ResultComponent, ResultComponentConfig, SimResultData } from './result_component';
