+++ ui/core/components/detailed_results/color_settings.ts
