+++ ui/core/components/detailed_results/player_damage.ts
@@ -1,5 +1,5 @@
-import { UnitMetrics, SimResult, SimResultFilter } from '../core/proto_utils/sim_result.js';
-import { maxIndex } from '../core/utils.js';
+import { UnitMetrics, SimResult, SimResultFilter } from '../../proto_utils/sim_result.js';
+import { maxIndex } from '../../utils.js';
 
 import { ColumnSortType, MetricsTable } from './metrics_table.js';
 import { ResultComponent, ResultComponentConfig, SimResultData } from './result_component.js';
