+++ ui/core/components/detailed_results/resource_metrics.ts
@@ -1,9 +1,7 @@
-import { ActionId } from '../core/proto_utils/action_id.js';
-import { ResourceMetrics, SimResult, SimResultFilter } from '../core/proto_utils/sim_result.js';
-import { ResourceType } from '../core/proto/api.js';
-import { resourceNames } from '../core/proto_utils/names.js';
-import { orderedResourceTypes } from '../core/proto_utils/utils.js';
-import { getEnumValues } from '../core/utils.js';
+import { ResourceMetrics, SimResult, SimResultFilter } from '../../proto_utils/sim_result.js';
+import { ResourceType } from '../../proto/api.js';
+import { resourceNames } from '../../proto_utils/names.js';
+import { orderedResourceTypes } from '../../proto_utils/utils.js';
 
 import { ColumnSortType, MetricsTable } from './metrics_table.js';
 import { ResultComponent, ResultComponentConfig, SimResultData } from './result_component.js';
