+++ ui/core/components/detailed_results/source_chart.ts
@@ -1,6 +1,6 @@
-import { Component } from '../core/components/component.js';
-import { SimResult, SimResultFilter, ActionMetrics } from '../core/proto_utils/sim_result.js';
-import { sum } from '../core/utils.js';
+import { Component } from '../../components/component.js';
+import { SimResult, SimResultFilter, ActionMetrics } from '../../proto_utils/sim_result.js';
+import { sum } from '../../utils.js';
 
 import { actionColors } from './color_settings.js';
