+++ ui/scss/core/components/detailed_results/_dps_histogram.scss
@@ -1,5 +1,5 @@
 .dps-histogram-root {
-	height: 30vh;
-	width: 90vw;
+	height: 25vh;
+	width: 90%;
 	margin: auto;
 }
