+++ ui/scss/core/components/detailed_results/_player_damage.scss
