+++ ui/scss/core/components/detailed_results/_resource_metrics.scss
