+++ ui/scss/core/components/detailed_results/_results_filter.scss
