+++ ui/scss/core/components/detailed_results/_source_chart.scss
