+++ ui/scss/core/components/detailed_results/_timeline.scss
@@ -270,7 +270,7 @@ $combo-points-color: #ffa07a;
 }
 .rotation-timeline-separator {
 	width: 100%;
-	border: 2px solid white;
+	border: 1px solid white;
 	margin: 10px 0;
 }
 .rotation-timeline-cast-icon {
