+++ ui/scss/core/components/detailed_results/_topline_results.scss
@@ -1,7 +1,7 @@
-@import "../../core/components/raid_sim_action";
+@import "../raid_sim_action";
 
 .topline-results-root {
-	margin-bottom: 1rem;
+	padding-bottom: 1rem;
 	display: flex;
 	justify-content: space-around;
 }
